Take a single input, then unzip it.
This is going to be a top-level binding and there's a compiler
requirement that we don't `unzip` or whatever in these (or, at least
it's an issue here). `fit` is going to be called from FFI and thus
the input to `pred` should reflect the *single* value that is returned
from `fit`.